requirements.txt
dowhy==0.11.1
clip==1.0
geatpy==2.7.0
pytorch-fid==0.3.0
scikit-learn==1.1.0
Cython==3.0.10
datasets==3.0.1
DateTime==5.5
diffusers==0.30.3
dill==0.3.8
et-xmlfile==1.1.0
filelock==3.16.1
flatbuffers==1.12
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.4.0
Gooey==1.0.8.1
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.1.0
huggingface-hub==0.26.0
humanfriendly==10.0
imageio==2.34.2
importlib_metadata==8.2.0
intel-cmplr-lib-ur==2024.2.0
intel-openmp==2024.2.0
joblib==1.4.2
keras==2.6.0
Keras-Preprocessing==1.1.2
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.3.4
mkl==2024.2.0
mkl-service==2.4.0
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.1
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
pandas==1.4.4
patsy==0.5.6
ply==3.11
propcache==0.2.0
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydot==3.0.1
pygame==2.5.2
pygtrie==2.5.0
pyreadline3==3.5.4
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.9.11
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scipy==1.10.1
six==1.15.0
sympy==1.13.1
tbb==2021.13.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.15.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.13
tokenizers==0.20.1
torch==1.12.1
torchaudio==0.12.1
torchfile==0.1.0
torchvision==0.13.1
tqdm==4.66.5
transformers==4.45.2
tzdata==2024.1
Werkzeug==3.0.3
wrapt==1.12.1
wxPython==4.2.1
xlrd==2.0.1
xxhash==3.5.0
yarl==1.15.2
zipp==3.19.2
zope.interface==6.4.post2
